move transport uri encoding into its own class
