+++ k9mail-library/src/main/java/com/fsck/k9/mail/TransportProvider.java
@@ -21,8 +21,7 @@ public class TransportProvider {
         String uri = storeConfig.getTransportUri();
         if (uri.startsWith("smtp")) {
             OAuth2TokenProvider oauth2TokenProvider = null;
-            return new SmtpTransport(storeConfig, new DefaultTrustedSocketFactory(context),
-                    oauth2TokenProvider);
+            return new SmtpTransport(storeConfig, new DefaultTrustedSocketFactory(context), oauth2TokenProvider);
         } else if (uri.startsWith("webdav")) {
             return new WebDavTransport(storeConfig);
         } else {
