+++ k9mail-library/src/main/java/com/fsck/k9/mail/transport/smtp/SmtpTransport.java
@@ -12,8 +12,6 @@ import java.net.InetSocketAddress;
 import java.net.Socket;
 import java.net.SocketAddress;
 import java.net.SocketException;
-import java.net.URI;
-import java.net.URISyntaxException;
 import java.security.GeneralSecurityException;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -36,7 +34,9 @@ import com.fsck.k9.mail.Message;
 import com.fsck.k9.mail.Message.RecipientType;
 import com.fsck.k9.mail.MessagingException;
 import com.fsck.k9.mail.ServerSettings;
+import com.fsck.k9.mail.ServerSettings.Type;
 import com.fsck.k9.mail.Transport;
+import com.fsck.k9.mail.TransportUris;
 import com.fsck.k9.mail.filter.Base64;
 import com.fsck.k9.mail.filter.EOLConvertingOutputStream;
 import com.fsck.k9.mail.filter.LineWrapOutputStream;
@@ -55,150 +55,11 @@ import static com.fsck.k9.mail.CertificateValidationException.Reason.MissingCapa
 import static com.fsck.k9.mail.K9MailLib.DEBUG_PROTOCOL_SMTP;
 
 public class SmtpTransport extends Transport {
-    public static final int SMTP_CONTINUE_REQUEST = 334;
-    public static final int SMTP_AUTHENTICATION_FAILURE_ERROR_CODE = 535;
+    private static final int SMTP_CONTINUE_REQUEST = 334;
+    private static final int SMTP_AUTHENTICATION_FAILURE_ERROR_CODE = 535;
 
     private TrustedSocketFactory mTrustedSocketFactory;
     private OAuth2TokenProvider oauthTokenProvider;
-
-    /**
-     * Decodes a SmtpTransport URI.
-     *
-     * NOTE: In contrast to ImapStore and Pop3Store, the authType is appended at the end!
-     *
-     * <p>Possible forms:</p>
-     * <pre>
-     * smtp://user:password:auth@server:port ConnectionSecurity.NONE
-     * smtp+tls+://user:password:auth@server:port ConnectionSecurity.STARTTLS_REQUIRED
-     * smtp+ssl+://user:password:auth@server:port ConnectionSecurity.SSL_TLS_REQUIRED
-     * </pre>
-     */
-    public static ServerSettings decodeUri(String uri) {
-        String host;
-        int port;
-        ConnectionSecurity connectionSecurity;
-        AuthType authType = null;
-        String username = null;
-        String password = null;
-        String clientCertificateAlias = null;
-
-        URI smtpUri;
-        try {
-            smtpUri = new URI(uri);
-        } catch (URISyntaxException use) {
-            throw new IllegalArgumentException("Invalid SmtpTransport URI", use);
-        }
-
-        String scheme = smtpUri.getScheme();
-        /*
-         * Currently available schemes are:
-         * smtp
-         * smtp+tls+
-         * smtp+ssl+
-         *
-         * The following are obsolete schemes that may be found in pre-existing
-         * settings from earlier versions or that may be found when imported. We
-         * continue to recognize them and re-map them appropriately:
-         * smtp+tls
-         * smtp+ssl
-         */
-        if (scheme.equals("smtp")) {
-            connectionSecurity = ConnectionSecurity.NONE;
-            port = ServerSettings.Type.SMTP.defaultPort;
-        } else if (scheme.startsWith("smtp+tls")) {
-            connectionSecurity = ConnectionSecurity.STARTTLS_REQUIRED;
-            port = ServerSettings.Type.SMTP.defaultPort;
-        } else if (scheme.startsWith("smtp+ssl")) {
-            connectionSecurity = ConnectionSecurity.SSL_TLS_REQUIRED;
-            port = ServerSettings.Type.SMTP.defaultTlsPort;
-        } else {
-            throw new IllegalArgumentException("Unsupported protocol (" + scheme + ")");
-        }
-
-        host = smtpUri.getHost();
-
-        if (smtpUri.getPort() != -1) {
-            port = smtpUri.getPort();
-        }
-
-        if (smtpUri.getUserInfo() != null) {
-            String[] userInfoParts = smtpUri.getUserInfo().split(":");
-            if (userInfoParts.length == 1) {
-                authType = AuthType.PLAIN;
-                username = decodeUtf8(userInfoParts[0]);
-            } else if (userInfoParts.length == 2) {
-                authType = AuthType.PLAIN;
-                username = decodeUtf8(userInfoParts[0]);
-                password = decodeUtf8(userInfoParts[1]);
-            } else if (userInfoParts.length == 3) {
-                // NOTE: In SmtpTransport URIs, the authType comes last!
-                authType = AuthType.valueOf(userInfoParts[2]);
-                username = decodeUtf8(userInfoParts[0]);
-                if (authType == AuthType.EXTERNAL) {
-                    clientCertificateAlias = decodeUtf8(userInfoParts[1]);
-                } else {
-                    password = decodeUtf8(userInfoParts[1]);
-                }
-            }
-        }
-
-        return new ServerSettings(ServerSettings.Type.SMTP, host, port, connectionSecurity,
-                authType, username, password, clientCertificateAlias);
-    }
-
-    /**
-     * Creates a SmtpTransport URI with the supplied settings.
-     *
-     * @param server
-     *         The {@link ServerSettings} object that holds the server settings.
-     *
-     * @return A SmtpTransport URI that holds the same information as the {@code server} parameter.
-     *
-     * @see com.fsck.k9.mail.store.StoreConfig#getTransportUri()
-     * @see SmtpTransport#decodeUri(String)
-     */
-    public static String createUri(ServerSettings server) {
-        String userEnc = (server.username != null) ?
-                encodeUtf8(server.username) : "";
-        String passwordEnc = (server.password != null) ?
-                encodeUtf8(server.password) : "";
-        String clientCertificateAliasEnc = (server.clientCertificateAlias != null) ?
-                encodeUtf8(server.clientCertificateAlias) : "";
-
-        String scheme;
-        switch (server.connectionSecurity) {
-            case SSL_TLS_REQUIRED:
-                scheme = "smtp+ssl+";
-                break;
-            case STARTTLS_REQUIRED:
-                scheme = "smtp+tls+";
-                break;
-            default:
-            case NONE:
-                scheme = "smtp";
-                break;
-        }
-
-        String userInfo;
-        AuthType authType = server.authenticationType;
-        // NOTE: authType is append at last item, in contrast to ImapStore and Pop3Store!
-        if (authType != null) {
-            if (AuthType.EXTERNAL == authType) {
-                userInfo = userEnc + ":" + clientCertificateAliasEnc + ":" + authType.name();
-            } else {
-                userInfo = userEnc + ":" + passwordEnc + ":" + authType.name();
-            }
-        } else {
-            userInfo = userEnc + ":" + passwordEnc;
-        }
-        try {
-            return new URI(scheme, userInfo, server.host, server.port, null, null,
-                    null).toString();
-        } catch (URISyntaxException e) {
-            throw new IllegalArgumentException("Can't create SmtpTransport URI", e);
-        }
-    }
-
 
     private String mHost;
     private int mPort;
@@ -219,9 +80,13 @@ public class SmtpTransport extends Transport {
             OAuth2TokenProvider oauth2TokenProvider) throws MessagingException {
         ServerSettings settings;
         try {
-            settings = decodeUri(storeConfig.getTransportUri());
+            settings = TransportUris.decodeTransportUri(storeConfig.getTransportUri());
         } catch (IllegalArgumentException e) {
             throw new MessagingException("Error while decoding transport URI", e);
+        }
+
+        if (settings.type == Type.SMTP) {
+            throw new IllegalArgumentException("Expected SMTP StoreConfig!");
         }
 
         mHost = settings.host;
